Moved OSS integration tests into integration test folder
