+++ storage/oss/package.json
@@ -23,10 +23,10 @@
     "lint-fix": "eslint --fix ./src/**/*.ts 1>&2 && sort-package-json",
     "spell-check": "cspell ** --config ./node_modules/@itwin/object-storage-common-config/cspell.json",
     "test": "",
-    "test:integration:backend": "node lib/test/backend/RunIntegrationTests.js",
+    "test:integration:backend": "node lib/test/integration/backend/RunIntegrationTests.js",
     "test:integration:frontend": "npm-run-all -p -r test:integration:frontend:startServer test:integration:frontend:runTests ",
-    "test:integration:frontend:runTests": "wait-on tcp:1223 -t 30000 && node lib/test/frontend/RunIntegrationTests.js",
-    "test:integration:frontend:startServer": "node lib/test/frontend/StartServer.js"
+    "test:integration:frontend:runTests": "wait-on tcp:1223 -t 30000 && node lib/test/integration/frontend/RunIntegrationTests.js",
+    "test:integration:frontend:startServer": "node lib/test/integration/frontend/StartServer.js"
   },
   "prettier": "@itwin/object-storage-common-config/prettier.json",
   "eslintConfig": {
