+++ storage/oss/src/test/integration/frontend/RunIntegrationTests.ts
@@ -12,10 +12,12 @@ const { bundledScriptFileName } = require(path.resolve(
   "..",
   "..",
   "..",
+  "..",
   "webpack.config.js"
 ));
 const bundledSetupScript = path.resolve(
   __dirname,
+  "..",
   "..",
   "..",
   "..",
