+++ storage/oss/webpack.config.js
@@ -15,7 +15,7 @@ const webpackConfig = {
     minimize: false
   },
   entry: {
-    app: path.resolve(__dirname, "lib", "test", "frontend", setupScriptFileName)
+    app: path.resolve(__dirname, "lib", "test", "integration", "frontend", setupScriptFileName)
   },
   output: {
     filename: bundledScriptFileName,
